Add withReadTimeout(Integer) method to NettyDockerCmdExecFactory.
This enhances our Netty support so that, like
JerseyDockerCmdExecFactory, we have a withReadTimeout(milliseconds)
method that configures the factory to implement a read timeout.
This allows clients to cope with recalcitrant docker daemons - if the
daemon stops responding then the client code will get a
SocketTimeoutException (much like it would if it was using Jersey
instead of Netty) instead of just hanging forever.